In History screen added Rv with PagingDataAdapter, but  I don't have date for rv items. I created HistoryPagingSource temporarily when Islom finished his task i change HistoryPagingSource class.

+++ app/src/main/java/uz/androbeck/virtualbank/data/dto/response/history/InComeAndOutComeResDto.kt
@@ -12,7 +12,7 @@ data class InComeAndOutComeResDto (
     @SerialName("to")
     val to: String?=null,
     @SerialName("amount")
-    val amount: Double=0.0,
+    val amount: Float=0f,
     @SerialName("time")
     val time: Long=0
 )
